IReservationRepository: for at JPA kan finde en reservation ud fra tlfnummer, skal vi bruge attributnavnene prÃæcist. Dvs. ikke custromerphone, men udelukkende phone som entitetsklassen

+++ src/main/java/org/example/adventurexp/dto/ActivityDTO.java
@@ -13,5 +13,5 @@ public class ActivityDTO {
     private Long id;
     private String name;
     private int minAge;
-    private double durationMinutes;
+    private int durationMinutes;
 }
